Handle windows with no _NET_WM_WINDOW_TYPE property
If the _NET_WM_WINDOW_TYPE property is not set for a window (like
URxvt), get_window_type() fails. Handle those windows as normal if they
don't have the WM_TRANSIENT_FOR property set. This is specified in the
Extended Window Manager Hints (EWMH) spec as follows:

  _NET_WM_WINDOW_TYPE_NORMAL indicates that this is a normal, top-level
  window, either managed or override-redirect. Managed windows with
  neither _NET_WM_WINDOW_TYPE nor WM_TRANSIENT_FOR set MUST be taken as
  this type.

  http://standards.freedesktop.org/wm-spec/latest/ar01s05.html

This change fixes the handling of rxvt-unicode windows.